Added functionality to nest to make the eggs searchable by the names of the actions used in the inputs.
This functionality works as follows:

(1) Whenever PlumedToHTML::get_html is called it adds the names of the actions that are found in the plumed input to a python set that is passed back to the calling code.
(2) An empty set is created when process_egg is called.  This empty set is then passed to PlumedToHTML::get_html every time a plumed input is called.  At the end of process_egg this set lists all the actions that have been used in the egg.
(3) The actions that are used within the egg are put in a column of the table that appears on the browse page.  This column is invisible but it is seachable.